refactor: move ordering types to `owidTypes.ts`
