+++ clientUtils/owidTypes.ts
@@ -15,6 +15,33 @@ export enum SortOrder {
     desc = "desc",
 }
 
+export enum SortBy {
+    entityName = "entityName",
+    dimension = "dimension",
+    total = "total",
+}
+
+interface SortByTotalConfig {
+    sortBy: SortBy.total
+    sortOrder: SortOrder
+}
+
+interface SortByEntityNameConfig {
+    sortBy: SortBy.entityName
+    sortOrder: SortOrder
+}
+
+export interface SortByDimensionConfig {
+    sortBy: SortBy.dimension
+    sortOrder: SortOrder
+    sortColumnSlug: string
+}
+
+export type SortConfig =
+    | SortByTotalConfig
+    | SortByEntityNameConfig
+    | SortByDimensionConfig
+
 export type Year = Integer
 export type Color = string
