+++ grapher/chart/ChartManager.ts
@@ -13,8 +13,7 @@ import { AxisConfigInterface } from "../axis/AxisConfigInterface"
 import { ColorSchemeName } from "../color/ColorConstants"
 import { EntityName } from "../../coreTable/OwidTableConstants"
 import { SelectionArray } from "../selection/SelectionArray"
-import { Annotation, ColumnSlug } from "../../clientUtils/owidTypes"
-import { SortConfig } from "../sort/SortState"
+import { Annotation, ColumnSlug, SortConfig } from "../../clientUtils/owidTypes"
 
 // The possible options common across our chart types. Not all of these apply to every chart type, so there is room to create a better type hierarchy.
